+++ assets/editor.asset.php
@@ -8,5 +8,5 @@ return array(
 		'wp-element',
 		'wp-i18n',
 	),
-	'version'      => '1.1.1',
+	'version'      => '1.1.2',
 );